Updated hero to use gsap module rather than script imports.
