+++ package.json
@@ -22,6 +22,7 @@
     "@mui/material": "^5.10.13",
     "@wordpress/block-library": "^7.13.0",
     "bootstrap": "5.1.3",
+    "gsap": "^3.11.3",
     "html-react-parser": "^3.0.4",
     "match-sorter": "^6.3.1",
     "next": "^12.0.7",
